[refactor🔨] : 상품 요청dto 중복 제거
