+++ src/main/java/shop/gaship/gashipshoppingmall/product/repository/custom/ProductRepositoryCustom.java
@@ -1,7 +1,7 @@
 package shop.gaship.gashipshoppingmall.product.repository.custom;
 
 import org.springframework.data.repository.NoRepositoryBean;
-import shop.gaship.gashipshoppingmall.product.dto.request.ProductRequestDto;
+import shop.gaship.gashipshoppingmall.product.dto.request.ProductRequestViewDto;
 import shop.gaship.gashipshoppingmall.product.dto.response.ProductAllInfoResponseDto;
 import shop.gaship.gashipshoppingmall.response.PageResponse;
 
@@ -21,5 +21,5 @@ public interface ProductRepositoryCustom {
      * @return list 조회된 상품들의 정보가들어있습니다.
      * @author 유호철
      */
-    PageResponse<ProductAllInfoResponseDto> findProduct(ProductRequestDto requestDto);
+    PageResponse<ProductAllInfoResponseDto> findProduct(ProductRequestViewDto requestDto);
 }
